requirements.txt
aiohttp>=3.9.0
click>=8.1.0
colorama>=0.4.6
pyyaml>=6.0.1
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
rich>=13.0.0
aiofiles>=23.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Dashboard
flask>=3.0.0
flask-cors>=4.0.0
